YAN-30 Добавил автоматическую генерацию волн + рефакторинг
