+++ packages/client/docs/gameEngine.md
@@ -208,5 +208,21 @@ return <canvas onClick={canvasClickHandler} ref={canvasRef} />
 - hearts - количество жизней при старте уровня
 - gameParams:
   - map - урл на карту уровня 
-  - enemyWaves - массив волн на уровень
-  - levelPlacements - массив мест, в которых допустимо строительство
+  - enemyWaves - массив волн (параметр опциональный, при его отсутствии волны сгенерируются автоматически)
+  - levelWayPoints - массив координат, по которым должны идти враги
+  - levelPlacements - массив мест, в которых допустимо строительство 
+
+### <a id="useStartGame"/> useAnimateCanvas
+Функция, которая отвечает за отрисовку каждого кадра.
+При вызове нужно передать следующие параметры:
+- context - контекст Canvas
+- background - картинку для фона
+- placementTiles - массив с плитками, на которых возможно строительство
+- buildingsRef - массив с существующими зданиями
+- enemiesRef массив с существующими врагами
+- mouseRef - позиция мыши
+- waves - массив с волнами
+- hearts - количество жизней
+- setHearts - функция для изменения количества жизней
+- onEnemyDefeated - функция, для обработки победы над врагом
+
