+++ packages/client/src/GameEngine/hooks/index.ts
@@ -3,3 +3,4 @@ export { default as usePlacementTiles } from './usePlacementTiles'
 export { default as useStartGame } from './useStartGame'
 export { default as useActiveTile } from './useActiveTile'
 export { default as useCanvasClickBuilder } from './useCanvasClickBuilder'
+export { default as useAnimateCanvas } from './useAnimateCanvas'
