+++ packages/client/src/GameEngine/hooks/useCanvasClickBuilder.ts
@@ -10,6 +10,8 @@ const useCanvasClickBuilder = (
 ) => {
   const coins = useSelector(getCoins)
   const dispatch = useDispatch()
+
+  // Строительство нового здания - для строительства необходимо, чтобы плитка была свободной
   const canvasClickBuilder = () => {
     const tile = activeTile.current
     if (tile && buildingsRef.current && !tile.occupied) {
@@ -20,6 +22,8 @@ const useCanvasClickBuilder = (
         },
       })
 
+      // Размещение здания стоит определённого количества монет, здание будет построено только если монет достаточно.
+      // После строительства баланс уменьшается.
       const isBuildingAvalible = coins >= newTower.price
       if (isBuildingAvalible) {
         buildingsRef.current.push(newTower)
