+++ packages/client/src/GameEngine/index.ts
@@ -1,7 +1,7 @@
-export * from './BaseClasses/Enemy'
-export * from './BaseClasses/Building'
-export * from './BaseClasses/Placement'
 export * from './hooks'
 export * from './utils'
 export * from './utils'
 export * from './baseConstants'
+export * from './BaseClasses/Enemy'
+export * from './BaseClasses/Building'
+export * from './BaseClasses/Placement'
