+++ packages/client/src/GameEngine/utils.ts
@@ -1,7 +1,8 @@
-import { IAnimateCanvasParams, ICreateEnemyOptions } from './types'
+import { ICreateEnemyOptions, IPosition } from './types'
 import { Enemy } from './BaseClasses/Enemy'
 import { collsOnMap } from './baseConstants'
 import { Building } from './BaseClasses/Building'
+import { Ork } from './GameElements'
 import { Projectile } from './BaseClasses/Projectile'
 
 export const createEnemies = (params: ICreateEnemyOptions): Enemy[] => {
@@ -51,90 +52,6 @@ export const calculateDistanceToBuilding = (
   return Math.hypot(xDistance, yDistance)
 }
 
-export const animateCanvasWrapper = (params: IAnimateCanvasParams) => {
-  const animationFrameCallback = () => animateCanvas(params)
-  const animationId = requestAnimationFrame(animationFrameCallback)
-  if (params.hearts.current === 0) {
-    cancelAnimationFrame(animationId)
-  }
-}
-
-const animateCanvas = (params: IAnimateCanvasParams) => {
-  const {
-    context,
-    background,
-    placementTiles,
-    enemiesRef,
-    mouseRef,
-    buildingsRef,
-    waves,
-    hearts,
-    setHearts,
-    onEnemyDefeated,
-  } = params
-  const buildings = buildingsRef.current
-  const enemies = enemiesRef.current
-
-  if (context) {
-    context.drawImage(background, 0, 0)
-    for (let i = enemies.length - 1; i >= 0; i--) {
-      const enemy = enemies[i]
-      enemy.update(context)
-      if (enemy.position.x > context.canvas.width) {
-        setHearts(hearts.current - 1)
-        enemies.splice(i, 1)
-      }
-    }
-
-    if (enemies.length === 0 && waves.length > 0) {
-      const nextWave = waves.shift()!
-      enemiesRef.current = createEnemies(nextWave)
-    }
-
-    placementTiles.forEach(tile => {
-      tile.update(context, mouseRef.current)
-    })
-
-    buildings?.forEach(building => {
-      building.update(context)
-      building.target = null
-
-      const validEnemies = enemies.filter(enemy => {
-        return (
-          calculateDistanceToBuilding(enemy, building)
-          < enemy.radius + building.radius
-        )
-      })
-      building.target = validEnemies[0]
-
-      for (let i = building.projectiles.length - 1; i >= 0; i--) {
-        const projectile = building.projectiles[i]
-        projectile.update(context)
-
-        const distance = calculateDistanceToProjectile(projectile)
-        if (distance < projectile.enemy.radius + projectile.radius) {
-          projectile.enemy.health -= projectile.power
-          if (projectile.enemy.health <= 0) {
-            const enemyIndex = enemies.findIndex(enemy => {
-              return projectile.enemy === enemy
-            })
-            if (enemyIndex > -1) {
-              onEnemyDefeated(
-                projectile.enemy.reward.coins,
-                projectile.enemy.reward.points
-              )
-              enemies.splice(enemyIndex, 1)
-            }
-          }
-          building.projectiles.splice(i, 1)
-        }
-      }
-    })
-
-    animateCanvasWrapper(params)
-  }
-}
-
 export const healthToPixels = (
   health: number,
   fullHealth: number,
@@ -146,4 +63,22 @@ export const healthToPixels = (
   const healthInPixels = (health / fullHealth) * fullHealthBarWidth
 
   return healthInPixels
+}
+
+export const generateSimpleWaves = (numberOfWaves: number, enemiesStep: number, wayPoints: IPosition[]): ICreateEnemyOptions[] => {
+  const waves: ICreateEnemyOptions[] = []
+
+  for (let i = 0; i < numberOfWaves; i++) {
+    const wave: ICreateEnemyOptions = {
+      levelPoints: wayPoints,
+      quantity: (i + 1) * enemiesStep,
+      gap: 100,
+      speed: 3 * (1 + ((i + 1) * enemiesStep)/50),
+      EnemyModel: Ork,
+    }
+
+    waves.push(wave)
+  }
+
+  return waves
 }
