+++ packages/client/src/components/Level/Level.tsx
@@ -1,15 +1,15 @@
 import React, { memo, useEffect, useRef } from 'react'
 import { FC } from 'react'
 import { useDispatch } from 'react-redux'
-import { newGame } from '../../../pages/Game/gameSlice'
-import { useCanvasClickBuilder, useStartGame } from '../../../GameEngine'
-import { enemyWavesLevel1, level1Placements } from './constants'
+import { newGame } from '../../pages/Game/gameSlice'
+import { useCanvasClickBuilder, useStartGame } from '../../GameEngine'
+import { level1Placements, level1WayPoints } from './constants'
 import map from './img/map.png'
-import { OrkTower } from '../../../GameEngine/GameElements'
+import { OrkTower } from '../../GameEngine/GameElements'
 
-const Level1: FC = () => {
+const Level: FC = () => {
   const dispatch = useDispatch()
-  const hearts = useRef<number>(2)
+  const hearts = useRef<number>(5)
 
   useEffect(() => {
     dispatch(newGame({ hearts: hearts.current, coins: 30 }))
@@ -18,8 +18,8 @@ const Level1: FC = () => {
   const { canvasRef, activeTileRef, buildingsRef } = useStartGame({
     hearts: hearts,
     gameParams: {
-      enemyWaves: enemyWavesLevel1,
       levelPlacements: level1Placements,
+      levelWayPoints: level1WayPoints,
       map,
     },
   })
@@ -33,4 +33,4 @@ const Level1: FC = () => {
   return <canvas onClick={canvasClickHandler} ref={canvasRef} />
 }
 
-export default memo(Level1)
+export default memo(Level)
