+++ packages/client/src/components/Level/img/map.png
