--- packages/client/src/components/Levels/index.ts
@@ -1,1 +1,0 @@
-export { default as Level1 } from './Level1/Level1'
